move utils script to a single Utils.py file
