+++ scripts/Utils.py
@@ -1,9 +1,27 @@
 import os
 import sys
-import urllib
+import winreg
+
 import requests
 import time
+import urllib
 
+from zipfile import ZipFile
+
+def GetSystemEnvVariable(name):
+    key = winreg.CreateKey(winreg.HKEY_LOCAL_MACHINE, r"SYSTEM\CurrentControlSet\Control\Session Manager\Environment")
+    try:
+        return winreg.QueryValueEx(key, name)[0]
+    except WindowsError:
+        return None
+    
+def GetUserEnvVariable(name):
+    key = winreg.CreateKey(winreg.HKEY_CURRENT_USER, r"Environment")
+    try:
+        return winreg.QueryValueEx(key, name)[0]
+    except WindowsError:
+        return None
+    
 def DownloadFromUrl(url, filepath):
     path = filepath
     filepath = os.path.abspath(filepath)
@@ -65,3 +83,45 @@ def DownloadFromUrl(url, filepath):
                 sys.stdout.write('\r[{}{}] {:.2f}% ({})     '.format('█' * done, '.' * (50-done), percentage, avgSpeedString))
                 sys.stdout.flush()
     sys.stdout.write('\n')
+
+def UnzipFile(filepath, deleteZipFile=True):
+    zipFilePath = os.path.abspath(filepath) # get full path of files
+    zipFileLocation = os.path.dirname(zipFilePath)
+
+    zipFileContent = dict()
+    zipFileContentSize = 0
+    with ZipFile(zipFilePath, 'r') as zipFileFolder:
+        for name in zipFileFolder.namelist():
+            zipFileContent[name] = zipFileFolder.getinfo(name).file_size
+        zipFileContentSize = sum(zipFileContent.values())
+        extractedContentSize = 0
+        startTime = time.time()
+        for zippedFileName, zippedFileSize in zipFileContent.items():
+            UnzippedFilePath = os.path.abspath(f"{zipFileLocation}/{zippedFileName}")
+            os.makedirs(os.path.dirname(UnzippedFilePath), exist_ok=True)
+            if os.path.isfile(UnzippedFilePath):
+                zipFileContentSize -= zippedFileSize
+            else:
+                zipFileFolder.extract(zippedFileName, path=zipFileLocation, pwd=None)
+                extractedContentSize += zippedFileSize
+            try:
+                done = int(50*extractedContentSize/zipFileContentSize)
+                percentage = (extractedContentSize / zipFileContentSize) * 100
+            except ZeroDivisionError:
+                done = 50
+                percentage = 100
+            elapsedTime = time.time() - startTime
+            try:
+                avgKBPerSecond = (extractedContentSize / 1024) / elapsedTime
+            except ZeroDivisionError:
+                avgKBPerSecond = 0.0
+            avgSpeedString = '{:.2f} KB/s'.format(avgKBPerSecond)
+            if (avgKBPerSecond > 1024):
+                avgMBPerSecond = avgKBPerSecond / 1024
+                avgSpeedString = '{:.2f} MB/s'.format(avgMBPerSecond)
+            sys.stdout.write('\r[{}{}] {:.2f}% ({})     '.format('█' * done, '.' * (50-done), percentage, avgSpeedString))
+            sys.stdout.flush()
+    sys.stdout.write('\n')
+
+    if deleteZipFile:
+        os.remove(zipFilePath) # delete zip file
